gh-96 - Rename to Typescript and fix imports
