+++ backend/package.json
@@ -2,18 +2,21 @@
   "name": "foosball-backend",
   "version": "0.2.0",
   "description": "The backend part of a Foosball Rating app.",
-  "main": "app.js",
+  "main": "app.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node  -r dotenv/config app.js"
+    "start": "ts-node --log-error -r dotenv/config app.ts"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/node": "^13.7.7",
     "dotenv": "^7.0.0",
     "express": "^4.16.4",
     "nedb": "^1.8.0",
     "pg": "^7.8.1",
-    "slackbots": "^1.2.0"
+    "slackbots": "^1.2.0",
+    "ts-node": "^8.4.1",
+    "typescript": "^3.6.3"
   }
 }
