+++ backend/rating/rating-calculator.ts
