+++ backend/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+      "module": "commonjs",
+      "target": "es6",
+      "sourceMap": true
+    },
+    "exclude": [
+      "node_modules"
+    ]
+  }
